:recycle: 레퍼런스 링크 API 로직 분리
