+++ frontend/src/components/PairRoom/ReferenceCard/ReferenceCard.tsx
@@ -1,16 +1,13 @@
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import { IoIosLink } from 'react-icons/io';
+
+import useReferenceLinks from '@/queries/PairRoom/useReferenceLinks';
 
 import Button from '@/components/common/Button/Button';
 import Input from '@/components/common/Input/Input';
 import { PairRoomCard } from '@/components/PairRoom/PairRoomCard';
-import Reference from '@/components/PairRoom/ReferenceCard/Reference';
-
-import { getReferenceLinks, addReferenceLink } from '@/apis/referenceLink';
+import ReferenceItem from '@/components/PairRoom/ReferenceCard/ReferenceItem';
 
 import useInput from '@/hooks/common/useInput';
-
-import { QUERY_KEYS } from '@/constants/queryKeys';
 
 import { theme } from '@/styles/theme';
 
@@ -23,22 +20,8 @@ interface ReferenceCardProps {
 }
 
 const ReferenceCard = ({ accessCode }: ReferenceCardProps) => {
-  const queryClient = useQueryClient();
-
-  const { data } = useQuery({
-    queryKey: [QUERY_KEYS.GET_REFERENCE_LINKS],
-    queryFn: () => getReferenceLinks({ accessCode }),
-  });
-
-  const { mutate } = useMutation({
-    mutationFn: addReferenceLink,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.GET_REFERENCE_LINKS] });
-    },
-    onError: (error) => alert(error.message),
-  });
-
-  const { inputValue, handleOnChange, resetInputValue } = useInput({ value: '', message: '', status: 'default' });
+  const { referenceLinks, addReferenceLink } = useReferenceLinks(accessCode);
+  const { inputValue, handleOnChange, resetInputValue } = useInput({ value: '', message: '', status: 'DEFAULT' });
 
   const buttonActive = inputValue.value !== '' && inputValue.status === 'default';
 
@@ -48,44 +31,42 @@ const ReferenceCard = ({ accessCode }: ReferenceCardProps) => {
   };
 
   const addReference = (url: string) => {
-    mutate({ accessCode, url });
+    addReferenceLink({ accessCode, url });
   };
 
   return (
-    <>
-      <PairRoomCard>
-        <PairRoomCard.Header icon={<IoIosLink color={theme.color.primary[500]} />} title="링크">
-          <S.ReferenceLinkForm onSubmit={handleSubmit}>
-            <Input
-              placeholder="링크를 입력해주세요."
-              value={inputValue.value}
-              status={inputValue.status as Status}
-              onChange={(event: React.ChangeEvent<HTMLInputElement>) => handleOnChange(event)}
-              label=""
-              message={inputValue.message}
-            />
-            <Button
-              disabled={!buttonActive}
-              css={S.buttonStyle}
-              color="SECONDARY"
-              rounded={true}
-              onClick={() => addReference(inputValue.value)}
-            >
-              링크 추가
-            </Button>
-          </S.ReferenceLinkForm>
-        </PairRoomCard.Header>
-        <S.ReferenceList>
-          {data ? (
-            data.map((data) => {
-              return <Reference key={data.id} link={data.url} />;
-            })
-          ) : (
-            <S.EmptyText>저장된 링크가 없습니다.</S.EmptyText>
-          )}
-        </S.ReferenceList>
-      </PairRoomCard>
-    </>
+    <PairRoomCard>
+      <PairRoomCard.Header icon={<IoIosLink color={theme.color.primary[500]} />} title="링크">
+        <S.ReferenceLinkForm onSubmit={handleSubmit}>
+          <Input
+            placeholder="링크를 입력해주세요."
+            value={inputValue.value}
+            status={inputValue.status as Status}
+            onChange={(event) => handleOnChange(event)}
+            label=""
+            message={inputValue.message}
+          />
+          <Button
+            disabled={!buttonActive}
+            css={S.buttonStyle}
+            color="SECONDARY"
+            rounded={true}
+            onClick={() => addReference(inputValue.value)}
+          >
+            링크 추가
+          </Button>
+        </S.ReferenceLinkForm>
+      </PairRoomCard.Header>
+      <S.ReferenceList>
+        {referenceLinks.length > 0 ? (
+          referenceLinks.map((link) => {
+            return <ReferenceItem key={link.id} link={link.url} />;
+          })
+        ) : (
+          <S.EmptyText>저장된 링크가 없습니다.</S.EmptyText>
+        )}
+      </S.ReferenceList>
+    </PairRoomCard>
   );
 };
