+++ frontend/src/queries/PairRoom/useReferenceLinks.ts
@@ -1,0 +1,24 @@
+import { useQuery, useQueryClient, useMutation } from '@tanstack/react-query';
+
+import { getReferenceLinks, addReferenceLink } from '@/apis/referenceLink';
+
+import { QUERY_KEYS } from '@/constants/queryKeys';
+
+const useReferenceLinks = (accessCode: string) => {
+  const queryClient = useQueryClient();
+
+  const { data } = useQuery({
+    queryKey: [QUERY_KEYS.GET_REFERENCE_LINKS],
+    queryFn: () => getReferenceLinks({ accessCode }),
+  });
+
+  const { mutate } = useMutation({
+    mutationFn: addReferenceLink,
+    onSuccess: () => queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.GET_REFERENCE_LINKS] }),
+    onError: (error) => alert(error.message),
+  });
+
+  return { referenceLinks: data || [], addReferenceLink: mutate };
+};
+
+export default useReferenceLinks;
